apps/ai-worker/pyproject.toml
[tool.poetry]
name = "ai-worker"
version = "0.1.0"
description = "AI Worker for OCR, RAG ingestion, and analysis - Investment Assistant Platform"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
python-multipart = "^0.0.20"
httpx = "^0.28.0"
openai = "^1.59.0"
pillow = "^11.0.0"
pdf2image = "^1.17.0"
psycopg = {extras = ["binary"], version = "^3.2.0"}
pgvector = "^0.3.0"
redis = "^5.2.0"
structlog = "^24.4.0"
pytesseract = "^0.3.13"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"
mypy = "^1.14.0"
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
